Iniciando modelagem Sequelize
